// src/lib/types/wms-api/createVasServiceDto.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * ElasticWMS API
 * Warehouse Management System API - Web, RF, and Integration endpoints
 * OpenAPI spec version: 1.0.0
 */
import type { CreateVasServiceDtoCategory } from './createVasServiceDtoCategory.js';

export interface CreateVasServiceDto {
  category?: CreateVasServiceDtoCategory;
  defaultRate?: number;
  description?: string;
  estimatedTimeMinutes?: number;
  isActive?: boolean;
  serviceCode: string;
  serviceName: string;
  uomId?: string;
}