// src/lib/types/wms-api/updateVasServiceDto.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * ElasticWMS API
 * Warehouse Management System API - Web, RF, and Integration endpoints
 * OpenAPI spec version: 1.0.0
 */
import type { UpdateVasServiceDtoCategory } from './updateVasServiceDtoCategory.js';

export interface UpdateVasServiceDto {
  category?: UpdateVasServiceDtoCategory;
  defaultRate?: number;
  description?: string;
  estimatedTimeMinutes?: number;
  isActive?: boolean;
  serviceName?: string;
  uomId?: string;
}