// src/lib/types/wms-api/setClientRateDto.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * ElasticWMS API
 * Warehouse Management System API - Web, RF, and Integration endpoints
 * OpenAPI spec version: 1.0.0
 */

export interface SetClientRateDto {
  clientId: string;
  currency?: string;
  effectiveDate?: string;
  expiryDate?: string;
  minCharge?: number;
  ratePerUnit: number;
  serviceId: string;
}