// src/lib/types/wms-api/createWorkstationDto.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * ElasticWMS API
 * Warehouse Management System API - Web, RF, and Integration endpoints
 * OpenAPI spec version: 1.0.0
 */
import type { CreateWorkstationDtoCapabilities } from './createWorkstationDtoCapabilities.js';
import type { CreateWorkstationDtoStationType } from './createWorkstationDtoStationType.js';

export interface CreateWorkstationDto {
  capabilities?: CreateWorkstationDtoCapabilities;
  facilityId: string;
  locationId?: string;
  stationType: CreateWorkstationDtoStationType;
  workstationCode: string;
  workstationName: string;
}