// src/lib/types/wms-api/updateWorkstationDto.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * ElasticWMS API
 * Warehouse Management System API - Web, RF, and Integration endpoints
 * OpenAPI spec version: 1.0.0
 */
import type { UpdateWorkstationDtoCapabilities } from './updateWorkstationDtoCapabilities.js';

export interface UpdateWorkstationDto {
  capabilities?: UpdateWorkstationDtoCapabilities;
  isActive?: boolean;
  isAvailable?: boolean;
  locationId?: string;
  workstationName?: string;
}