// src/lib/types/wms-api/vasCatalogWebControllerListWorkstationsParams.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * ElasticWMS API
 * Warehouse Management System API - Web, RF, and Integration endpoints
 * OpenAPI spec version: 1.0.0
 */

export type VasCatalogWebControllerListWorkstationsParams = {
facilityId: string;
stationType: string;
isAvailable: string;
};